+++ src/main/java/cl/desagen/chilquinta/repositories/EstadosRepository.java
@@ -3,6 +3,6 @@ package cl.desagen.chilquinta.repositories;
 import cl.desagen.chilquinta.entities.EstadosEntity;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-public interface EstadosRepository extends PagingAndSortingRepository<EstadosEntity, Long> {
+public interface EstadosRepository extends PagingAndSortingRepository<EstadosEntity, Integer> {
 
 }
